--- a/crates/core/src/ctx/context.rs
+++ b/crates/core/src/ctx/context.rs
@@ -18,8 +18,10 @@ use async_channel::Sender;
 use std::borrow::Cow;
 use std::collections::HashMap;
 use std::fmt::{self, Debug};
+use std::future::Future;
 #[cfg(storage)]
 use std::path::PathBuf;
+use std::pin::Pin;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
@@ -359,40 +361,46 @@ impl MutableContext {
 	/// We may not want to check for the deadline on every call.
 	/// An iteration loop may want to check it every 10 or 100 calls.
 	/// Eg.: ctx.done(count % 100 == 0)
-	pub(crate) fn done(&self, deep_check: bool) -> Result<Option<Reason>, Error> {
-		match self.deadline {
-			Some(deadline) if deep_check && deadline <= Instant::now() => {
-				Ok(Some(Reason::Timedout))
-			}
-			_ if self.cancelled.load(Ordering::Relaxed) => Ok(Some(Reason::Canceled)),
-			_ => {
-				if deep_check && ALLOC.is_beyond_threshold() {
-					return Err(Error::QueryBeyondMemoryThreshold);
+	pub(crate) fn done<'a>(
+		&'a self,
+		deep_check: bool,
+	) -> Pin<Box<dyn Future<Output = Result<Option<Reason>, Error>> + Send + 'a>> {
+		Box::pin(async move {
+			yield_now!();
+			match self.deadline {
+				Some(deadline) if deep_check && deadline <= Instant::now() => {
+					Ok(Some(Reason::Timedout))
 				}
-				match &self.parent {
-					Some(ctx) => ctx.done(deep_check),
-					_ => Ok(None),
+				_ if self.cancelled.load(Ordering::Relaxed) => Ok(Some(Reason::Canceled)),
+				_ => {
+					if deep_check && ALLOC.is_beyond_threshold() {
+						return Err(Error::QueryBeyondMemoryThreshold);
+					}
+					match &self.parent {
+						Some(ctx) => ctx.done(deep_check).await,
+						_ => Ok(None),
+					}
 				}
 			}
-		}
+		})
 	}
 
 	/// Check if the context is ok to continue.
-	pub(crate) fn is_ok(&self, deep_check: bool) -> Result<bool, Error> {
-		Ok(self.done(deep_check)?.is_none())
+	pub(crate) async fn is_ok(&self, deep_check: bool) -> Result<bool, Error> {
+		Ok(self.done(deep_check).await?.is_none())
 	}
 
 	/// Check if there is some reason to stop processing the current query.
 	///
 	/// Returns true when the query is canceled or if check_deadline is true when the query
 	/// deadline is met.
-	pub(crate) fn is_done(&self, deep_check: bool) -> Result<bool, Error> {
-		Ok(self.done(deep_check)?.is_some())
+	pub(crate) async fn is_done(&self, deep_check: bool) -> Result<bool, Error> {
+		Ok(self.done(deep_check).await?.is_some())
 	}
 
 	/// Check if the context is not ok to continue, because it timed out.
-	pub(crate) fn is_timedout(&self) -> Result<bool, Error> {
-		Ok(matches!(self.done(true)?, Some(Reason::Timedout)))
+	pub(crate) async fn is_timedout(&self) -> Result<bool, Error> {
+		Ok(matches!(self.done(true).await?, Some(Reason::Timedout)))
 	}
 
 	#[cfg(storage)]