--- a/crates/core/src/dbs/processor.rs
+++ b/crates/core/src/dbs/processor.rs
@@ -462,7 +462,7 @@ pub(super) trait Collector {
 		opt: &Options,
 		iterable: Iterable,
 	) -> Result<(), Error> {
-		if ctx.is_ok(true)? {
+		if ctx.is_ok(true).await? {
 			match iterable {
 				Iterable::Value(v) => {
 					if v.is_some() {
@@ -540,7 +540,7 @@ pub(super) trait Collector {
 		let mut skipped = 0;
 		let mut last_key = vec![];
 		while let Some(res) = stream.next().await {
-			if ctx.is_done(skipped % 100 == 0)? {
+			if ctx.is_done(skipped % 100 == 0).await? {
 				break;
 			}
 			last_key = res?;
@@ -594,7 +594,7 @@ pub(super) trait Collector {
 		let mut count = 0;
 		while let Some(res) = stream.next().await {
 			// Check if the context is finished
-			if ctx.is_done(count % 100 == 0)? {
+			if ctx.is_done(count % 100 == 0).await? {
 				break;
 			}
 			// Parse the data from the store
@@ -635,7 +635,7 @@ pub(super) trait Collector {
 		let mut count = 0;
 		while let Some(res) = stream.next().await {
 			// Check if the context is finished
-			if ctx.is_done(count % 100 == 0)? {
+			if ctx.is_done(count % 100 == 0).await? {
 				break;
 			}
 			// Parse the data from the store
@@ -728,7 +728,7 @@ pub(super) trait Collector {
 		let mut count = 0;
 		while let Some(res) = stream.next().await {
 			// Check if the context is finished
-			if ctx.is_done(count % 100 == 0)? {
+			if ctx.is_done(count % 100 == 0).await? {
 				break;
 			}
 			// Parse the data from the store
@@ -767,7 +767,7 @@ pub(super) trait Collector {
 		let mut count = 0;
 		while let Some(res) = stream.next().await {
 			// Check if the context is finished
-			if ctx.is_done(count % 100 == 0)? {
+			if ctx.is_done(count % 100 == 0).await? {
 				break;
 			}
 			// Parse the data from the store
@@ -878,7 +878,7 @@ pub(super) trait Collector {
 			let mut count = 0;
 			while let Some(res) = stream.next().await {
 				// Check if the context is finished
-				if ctx.is_done(count % 100 == 0)? {
+				if ctx.is_done(count % 100 == 0).await? {
 					break;
 				}
 				// Parse the key from the result
@@ -936,14 +936,14 @@ pub(super) trait Collector {
 		txn: &Transaction,
 		mut iterator: ThingIterator,
 	) -> Result<(), Error> {
-		while !ctx.is_done(true)? {
+		while !ctx.is_done(true).await? {
 			let records: Vec<IndexItemRecord> =
 				iterator.next_batch(ctx, txn, *NORMAL_FETCH_SIZE).await?;
 			if records.is_empty() {
 				break;
 			}
 			for (c, r) in records.into_iter().enumerate() {
-				if ctx.is_done(c % 100 == 0)? {
+				if ctx.is_done(c % 100 == 0).await? {
 					break;
 				}
 				self.collect(Collected::IndexItemKey(r)).await?;
@@ -958,14 +958,14 @@ pub(super) trait Collector {
 		txn: &Transaction,
 		mut iterator: ThingIterator,
 	) -> Result<(), Error> {
-		while !ctx.is_done(true)? {
+		while !ctx.is_done(true).await? {
 			let records: Vec<IndexItemRecord> =
 				iterator.next_batch(ctx, txn, *NORMAL_FETCH_SIZE).await?;
 			if records.is_empty() {
 				break;
 			}
 			for (c, r) in records.into_iter().enumerate() {
-				if ctx.is_done(c % 100 == 0)? {
+				if ctx.is_done(c % 100 == 0).await? {
 					break;
 				}
 				self.collect(Collected::IndexItem(r)).await?;
@@ -981,7 +981,7 @@ pub(super) trait Collector {
 		mut iterator: ThingIterator,
 	) -> Result<(), Error> {
 		let mut total_count = 0;
-		while !ctx.is_done(true)? {
+		while !ctx.is_done(true).await? {
 			let count = iterator.next_count(ctx, txn, *NORMAL_FETCH_SIZE).await?;
 			if count == 0 {
 				break;