--- a/crates/core/src/sql/statements/create.rs
+++ b/crates/core/src/sql/statements/create.rs
@@ -81,7 +81,7 @@ impl CreateStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results