--- a/crates/core/src/sql/statements/delete.rs
+++ b/crates/core/src/sql/statements/delete.rs
@@ -71,7 +71,7 @@ impl DeleteStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results