--- a/crates/core/src/sql/statements/insert.rs
+++ b/crates/core/src/sql/statements/insert.rs
@@ -126,7 +126,7 @@ impl InsertStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results