--- a/crates/core/src/sql/statements/relate.rs
+++ b/crates/core/src/sql/statements/relate.rs
@@ -173,7 +173,7 @@ impl RelateStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results