--- a/crates/core/src/sql/statements/select.rs
+++ b/crates/core/src/sql/statements/select.rs
@@ -136,7 +136,7 @@ impl SelectStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, &opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results