--- a/crates/core/src/sql/statements/update.rs
+++ b/crates/core/src/sql/statements/update.rs
@@ -72,7 +72,7 @@ impl UpdateStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results