--- a/crates/core/src/sql/statements/upsert.rs
+++ b/crates/core/src/sql/statements/upsert.rs
@@ -71,7 +71,7 @@ impl UpsertStatement {
 		// Process the statement
 		let res = i.output(stk, &ctx, opt, &stm, RecordStrategy::KeysAndValues).await?;
 		// Catch statement timeout
-		if ctx.is_timedout()? {
+		if ctx.is_timedout().await? {
 			return Err(Error::QueryTimedout);
 		}
 		// Output the results