Thank you for submitting this pull request. We really appreciate you spending the time to work on SurrealDB. 🚀 🎉 

## What is the motivation?

We want to reduce async runtime latency by improving cooperative yielding.

## What does this change do?

Identify long-duration loops and iterations and include yielding calls.

## What is your testing strategy?

Github action.
Covered by current test.

CRUD-BENCH has been run to assess the performance impact.
No noticeable decrease in performance has been observed.

Main branch:

```
Create took 49s 811ms
Read took 16s 281ms
Update took 1m 2s
Scan::count_all took 1m 28s
Scan::limit_id took 25ms 847µs
Scan::limit_all took 39ms 796µs
Scan::limit_count took 26ms 512µs
Scan::limit_start_id took 54ms 81µs
Scan::limit_start_all took 70ms 59µs
Scan::limit_start_count took 55ms 928µs
Delete took 46s 791ms
```

This branch:

```
Create took 47s 713ms
Read took 16s 358ms
Update took 58s 917ms
Scan::count_all took 1m 29s
Scan::limit_id took 31ms 686µs
Scan::limit_all took 41ms 392µs
Scan::limit_count took 35ms 908µs
Scan::limit_start_id took 63ms 672µs
Scan::limit_start_all took 69ms 345µs
Scan::limit_start_count took 68ms 252µs
Delete took 44s 454ms
```



## Is this related to any issues?

#5744
#5617 

## Does this change need documentation?

<!-- If this pull request requires changes, updates, or improvements to the documentation, then add a corresponding issue on the https://github.com/surrealdb/docs.surrealdb.com repository, and link to it here. -->

- [x] No documentation needed

## Have you read the Contributing Guidelines?

<!-- All pull requests require that the contributing guidelines have been read and agreed to. -->

- [x] I have read the [Contributing Guidelines](https://github.com/surrealdb/surrealdb/blob/main/CONTRIBUTING.md)